requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.1
django-filter==24.1
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
